+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -8,6 +8,7 @@ unreleased
 
 * Add support for incremental blob copy
 * Add support for large block blob upload
+* Change block size to 100MB when file to upload is larger than 200GB
 
 2.0.2 (2017-04-03)
 ++++++++++++++++++