+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ===============
 
-unreleased
+2.0.3 (2017-04-17)
 ++++++++++++++++++
 
 * Add support for incremental blob copy